fix tuning table nvram data size

+++ src/main.h
@@ -60,7 +60,7 @@ typedef const struct {
 	midi_arp_state_t midi_arp_state;
 	tt_state_t tt_state;
 	uint8_t scale[16][8];
-	uint8_t tuning_table[4][120];
+	uint16_t tuning_table[4][120];
 } nvram_data_t;
 
 extern nvram_data_t f;
